+++ src/pages/admin/postManagement/PostManagement.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef, useContext } from 'react';
-import { getPosts } from '../../api/project_api';
+import { getPosts } from '../../../api/project_api';
 import dayjs from 'dayjs';
-import { AuthContext } from '../../context/AuthContext';
+import { AuthContext } from '../../../context/AuthContext';
 import { useNavigate } from 'react-router-dom';
 
 export default function PostManagement() {
